chore: move peer id generator to peer package
