+++ client/daemon/proxy/proxy.go
@@ -105,7 +105,7 @@ type Proxy struct {
 	// dumpHTTPContent indicates to dump http request header and response header
 	dumpHTTPContent bool
 
-	peerIDGenerator transport.PeerIDGenerator
+	peerIDGenerator peer.IDGenerator
 }
 
 // Option is a functional option for configuring the proxy
@@ -115,7 +115,14 @@ type Option func(p *Proxy) *Proxy
 func WithPeerHost(peerHost *scheduler.PeerHost) Option {
 	return func(p *Proxy) *Proxy {
 		p.peerHost = peerHost
-		p.peerIDGenerator = transport.NewPeerIDGenerator(peerHost.Ip)
+		return p
+	}
+}
+
+// WithPeerIDGenerator sets the *transport.PeerIDGenerator
+func WithPeerIDGenerator(peerIDGenerator peer.IDGenerator) Option {
+	return func(p *Proxy) *Proxy {
+		p.peerIDGenerator = peerIDGenerator
 		return p
 	}
 }
@@ -255,9 +262,7 @@ func (proxy *Proxy) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	span.SetAttributes(semconv.HTTPHostKey.String(r.Host))
 	span.SetAttributes(semconv.HTTPURLKey.String(r.URL.String()))
 	span.SetAttributes(semconv.HTTPMethodKey.String(r.Method))
-	defer func() {
-		span.End()
-	}()
+	defer span.End()
 
 	// update ctx to transfer trace id
 	r = r.WithContext(ctx)
@@ -267,6 +272,7 @@ func (proxy *Proxy) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		user, pass, ok := proxyBasicAuth(r)
 		if !ok {
 			status := http.StatusProxyAuthRequired
+			span.SetAttributes(semconv.HTTPStatusCodeKey.Int(status))
 			http.Error(w, http.StatusText(status), status)
 			logger.Debugf("empty auth info: %s, url：%s", r.Host, r.URL.String())
 			return
@@ -274,6 +280,7 @@ func (proxy *Proxy) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		// TODO dynamic auth config via manager
 		if user != proxy.basicAuth.Username || pass != proxy.basicAuth.Password {
 			status := http.StatusUnauthorized
+			span.SetAttributes(semconv.HTTPStatusCodeKey.Int(status))
 			http.Error(w, http.StatusText(status), status)
 			logger.Debugf("mismatch auth info (%s/%s): %s, url：%s", user, pass, r.Host, r.URL.String())
 			return
@@ -286,6 +293,7 @@ func (proxy *Proxy) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	// check whiteList
 	if !directRequest && !proxy.checkWhiteList(r) {
 		status := http.StatusUnauthorized
+		span.SetAttributes(semconv.HTTPStatusCodeKey.Int(status))
 		http.Error(w, http.StatusText(status), status)
 		logger.Debugf("not in whitelist: %s, url：%s", r.Host, r.URL.String())
 		return
